move template sections to own file to avoid costly imports
