+++ scripts/changelog/changelogTemplateSections.ts
@@ -1,0 +1,6 @@
+export const changelogTemplateSections = {
+  androidUserFacingChanges: "Android user-facing changes",
+  crossPlatformUserFacingChanges: "Cross-platform user-facing changes",
+  devChanges: "Dev changes",
+  iOSUserFacingChanges: "iOS user-facing changes",
+}
